Allow mapping to have an empty string as name
The mapping name check in addMapping was previously for truthiness which
meant that an empty string would evaluate to false. This causes an error
in the _serializeMappings function when a mapping has an empty string
for its name because that check is for != null and thus an entry is
expected in _names and it does not exist.