Add a column argument to allGeneratedPositionsFor.
Some source maps define multiple mappings for a single original column. We need
a way to obtain all mappings for a given column, instead of just the first one.
This patch refactors allGeneratedPositionsFor so that it returns all mappings
for a single line if no column argument is passed (not necessarily the line that
was requested, due to fuzzy search), and all mappings for a single line and
column otherwise.