Skipping mappings that exceed line count in generated code
This is a fix for #174. In `fromStringWithSourceMap` we should not
attempt to create mappings for the SourceNode when there are mappings
for lines that exceed the number of lines in the generated code. This
can happen in scenarios where a sourcemap has been transformed a number
of times and new mappings are added each time. There will be "dead
mappings" that refer to generatedLines that no longer exceed in the
generated code. We can safely ignore these. Fixes #174