Fix util.relative for the case where the path is above the root.
Currently, util.relative simply checks if the root is a prefix of the path. This
works fine when the path is below the root, but won't work if the path is above
the root.

To fix this, we should remove components from the root one by one, until either
we find a prefix that fits, or we run out of components to match.