Move off of dryice and to webpack for builds.
We were pretty much the last dryice user, and it isn't worth being on such a
minority build tool when we don't really want to maintain it
ourselves. Additionally, this lets us stop using amdefine and the AMD format,
which was an alright choice in 2011 but has increasinly shown its age. Regular
Common JS modules are used now, which should pave the way for better tree
shaking and unused submodule removal tooling. Note that in order to avoid
ruining `git blame`, this leaves each module indented by two spaces as it was
from the AMD `define` function, and wraps a block around them so that editors
don't freak out too much.