Remove blocks around each module
When switching from AMD style modules to common JS modules, we wanted to keep
the diff as small as possible, and the solution was to wrap each module in a {}
block. This preserved blame and everyone was happy. But it turns out that
SpiderMonkey won't hoist functions that are defined in blocks, so we have a
really subtle foot-gun attached to a time bomb (that has already gone off once:
see #233).

I'd much prefer not to have any weaponized code aimed at my foot, so I'm trading
our blame preservation for the lack of time bombs and foot guns.

Here is the diff ignoring white-space: https://pastebin.mozilla.org/8868794