Fixed "readFileSync" example
The method `fs.readFileSync` returns a Buffer when no encoding is specified (https://nodejs.org/api/fs.html#fs_fs_readfilesync_file_options) and this causes the `SourceMapConsumer` to fail because it expects a string to execute `JSON.parse`. Therefore an encoding needs to be specified, otherwise SourceMapsConsumer fails with `Error: "version" is a required argument`.