Fix which "URL" to use
If the global "URL" is a function, use it. This should be the case in
browser environments.

In [node] environment, `require('url').URL`, which follows WHATWG URL
Standard, was added in v7, and `URL` became available in global object
since v10.0. To support the version before v10, we need to check the
global object, and then fallback to `require('url').URL`.

However, note one thing: the version packed in by webpack-v3 ([npm-url])
only supports the "legacy version of Url API", which is not compatible
with "WHATWG URL Standard", must not be used as a replacement for `URL`.

[node]: https://nodejs.org/dist/latest-v10.x/docs/api/url.html
[npm-url]: https://www.npmjs.com/package/url